requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
google-cloud-storage==2.14.0
google-generativeai==0.3.2
Pillow==11.2.1 --only-binary :all:
PyMuPDF==1.26.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Werkzeug==3.1.3
tqdm==4.66.2
numpy==2.2.0
scikit-learn==1.5.2
celery==5.5.3
redis==6.2.0

